WIP refactoring - renamings
