+++ src/roman_numerals/handler.py
@@ -1,42 +1,31 @@
 from abc import ABC, abstractmethod
-from typing import List
 
-from src.roman_numerals.numeral_gateway import NumeralDescriptor
+from src.roman_numerals.decimal_system import DecimalSystem
 
 
 class Handler(ABC):
     @abstractmethod
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         pass
 
     @abstractmethod
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         pass
 
 
-class ArabicToRomanNumeralService:
-    def __init__(self, handlers: List[Handler]):
-        self.handlers = handlers
-
-    def convert(self, request: NumeralDescriptor, mapping) -> str:
-        for handler in self.handlers:
-            if handler.can_handle(request):
-                return handler.handle(request, mapping)
-
-
 class ExactMatchHandler(Handler):
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         res = request.get_closest_greater_distance() == 0
         if res:
             print("ExactMatchHandler")
         return res
 
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         return mapping.get_roman_numeral(request.value)
 
 
 class BeforeExactMatchHandler(Handler):
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         closest_greater_number_distance = request.get_closest_greater_distance()
         res = (
             closest_greater_number_distance
@@ -46,7 +35,7 @@ class BeforeExactMatchHandler(Handler):
             print("BeforeExactMatchHandler")
         return res
 
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         second_closest_smaller_number = request.get_second_smaller_number()
         closest_greater_number = request.get_closest_greater_number()
         return mapping.get_roman_numeral(
@@ -55,14 +44,14 @@ class BeforeExactMatchHandler(Handler):
 
 
 class AfterSmallerMatchHandler(Handler):
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         closest_smaller_number = request.get_closest_smaller_number()
         res = request.value % closest_smaller_number == 0
         if res:
             print("AfterSmallerMatchHandler")
         return res
 
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         closest_smaller_number = request.get_closest_smaller_number()
         return mapping.get_roman_numeral(closest_smaller_number) * round(
             request.value / closest_smaller_number
@@ -70,13 +59,13 @@ class AfterSmallerMatchHandler(Handler):
 
 
 class BeforeGreaterMatchHandler(Handler):
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         res = True
         if res:
             print("BeforeGreaterMatchHandler")
         return res
 
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         closest_smaller_number = request.get_closest_smaller_number()
         second_closest_smaller_number = request.get_second_smaller_number()
         roman_numeral_substring = mapping.get_roman_numeral(
@@ -90,8 +79,8 @@ class BeforeGreaterMatchHandler(Handler):
 
 
 class ErrorHandler(Handler):
-    def can_handle(self, request: NumeralDescriptor) -> bool:
+    def can_handle(self, request: DecimalSystem) -> bool:
         return True
 
-    def handle(self, request: NumeralDescriptor, mapping) -> str:
+    def handle(self, request: DecimalSystem, mapping) -> str:
         return f"Unsupported request {request.value}"
